Remove step that does not instruct the user to take some action
The step with the text "Now you are in the Test tab." (i.e., substep 2 of **Step 6: Test the service**) does not instruct the user to take any action. Therefore, that text is moved to the previous step (substep 1) so that it is depicted as a result of performing substep 1.