Refactor & fix invalid GET sent with empty body
When testing the generated curl an empty body(`body: {}`) is being sent with the GET call by default
which causes the backend to consider the request as invalid. Removed the body for now since we are only
making GET requests in all specs.

Also refactored the test to make the scenarios being tested more clear.