Fix data cleanup to only remove new test data
Data cleanup was meant to be done for projects starting with the prefix `automationtestproject` but skip those
with the prefix `testproject`. The logic was excluding the new name format as well since the old format is a
substring of the new. Fixed the condition to match the starting string of the project name instead of checking
if the string is contained in the name.