Workaround for https://github.com/wso2-enterprise/choreo/issues/29369
Due to the issue mantioned in https://github.com/wso2-enterprise/choreo/issues/29369 it may take a
siginificnt time for APIs to be available on the DP after successful deployment. The increased wait
time and retries should help to mitigate this issue and increase the chances of tests passing when
this issue in encountered.